refactor: move conversion-related files to util namespace
