+++ sese/internal/unix/system/SharedMemory.cpp
@@ -14,7 +14,7 @@
 
 #include <sese/system/SharedMemory.h>
 #include <sese/io/ByteBuilder.h>
-#include <sese/convert/MD5Util.h>
+#include <sese/util/MD5Util.h>
 
 using namespace sese::system;
