+++ sese/record/ConsoleAppender.cpp
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include "sese/record/ConsoleAppender.h"
-// #include "sese/convert/EncodingConverter.h"
+// #include "sese/util/EncodingConverter.h"
 
 #include <cstdio>
