+++ sese/test/TestEncodingConverter.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <sese/convert/EncodingConverter.h>
+#include <sese/util/EncodingConverter.h>
 
 #include <gtest/gtest.h>
