+++ sese/util/Base64Converter.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "sese/convert/Base64Converter.h"
+#include "sese/util/Base64Converter.h"
 #include "sese/util/Endian.h"
 
 #include <cmath>
