+++ sese/util/Compressor.h
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 /**
- * \file Decompressor.h
- * \brief Decompressor class
+ * \file Compressor.h
+ * \brief Compressor class
  * \version 0.1
  * \date March 4, 2023
  * \author kaoru
@@ -22,36 +22,38 @@
 
 #pragma once
 
-#include "sese/convert/ZlibConfig.h"
+#include "sese/util/ZlibConfig.h"
 #include "sese/io/OutputStream.h"
 
 namespace sese {
 
-/// Decompressor class
-class Decompressor {
+/// Compressor class
+class Compressor {
 public:
     using OutputStream = sese::io::OutputStream;
 
-    /// Create a decompressor
-    /// \param type Compression format
+    /// Create a compressor
+    /// \param type Compression format - GZIP option is not recommended here
+    /// \param level Compression level
     /// \param buffer_size Set internal buffer size
-    explicit Decompressor(CompressionType type, size_t buffer_size = ZLIB_CHUNK_SIZE);
+    explicit Compressor(CompressionType type, size_t level, size_t buffer_size = ZLIB_CHUNK_SIZE);
 
-    virtual ~Decompressor();
+    virtual ~Compressor();
 
-    /// Set the buffer to be decompressed
-    /// \param input Buffer to be decompressed
+    /// Set the buffer to be compressed
+    /// \param input Buffer to be compressed
     /// \param input_size Size of this buffer
     void input(const void *input, unsigned int input_size);
 
-    /// Perform decompression
-    /// \param out Decompressed stream
-    /// \retval Z_OK (0) Current buffer block decompressed successfully
+    /// Perform compression
+    /// \param out Output stream for compressed data
+    /// \retval Z_OK (0) Current buffer block compressed successfully
     /// \retval Z_STREAM_ERROR (-2) Other errors
     /// \retval Z_BUF_ERROR (-5) Output stream capacity insufficient
-    int inflate(OutputStream *out);
+    int deflate(OutputStream *out);
 
     /// Reset z_stream object
+    /// \return Result
     int reset();
 
     /// Size of the currently processed input buffer
