+++ sese/util/EncodingConverter.h
