+++ sese/util/PercentConverter.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "sese/convert/PercentConverter.h"
+#include "sese/util/PercentConverter.h"
 #include "sese/text/StringBuilder.h"
 
 using sese::PercentConverter;
